Address Claude review feedback
- Centralize getSpatialExtensionPath in @geolibre/core (runtime-env.ts) and
  export it, replacing the line-for-line duplicate that lived in the plugin.
  Both consumers (desktop duckdb loader via the spatial-extension-config
  re-export, and the maplibre-vector plugin) now share one implementation, so
  the VITE_DUCKDB_SPATIAL_EXTENSION_PATH logic cannot silently diverge.
- Drop the now-redundant tests/vector-spatial-extension-path.test.ts; the same
  logic is covered by tests/spatial-extension-config.test.ts.
