+++ apps/geolibre-desktop/src/lib/spatial-extension-config.ts
@@ -1,9 +1,5 @@
-import { getRuntimeEnvironment } from "@geolibre/core";
-
-export function getSpatialExtensionPath(
-  env?: Record<string, string | undefined>,
-): string | undefined {
-  const runtimeEnv = env ?? getRuntimeEnvironment();
-  const value = runtimeEnv.VITE_DUCKDB_SPATIAL_EXTENSION_PATH;
-  return value && value.trim() ? value.trim() : undefined;
-}
+// The shared implementation lives in @geolibre/core so the desktop loader and
+// the maplibre-gl-vector plugin resolve VITE_DUCKDB_SPATIAL_EXTENSION_PATH the
+// same way. Re-exported here to keep this module path stable for existing
+// importers (sql-workspace, duckdb-vector-loader, ...).
+export { getSpatialExtensionPath } from "@geolibre/core";
