+++ packages/core/src/index.ts
@@ -71,4 +71,4 @@ export {
   type GeocodeRequest,
   type ReverseGeocodeDisplay,
 } from "./geocoding";
-export { getRuntimeEnvironment } from "./runtime-env";
+export { getRuntimeEnvironment, getSpatialExtensionPath } from "./runtime-env";
